Move 3d map control into its own folder, remove unused code.
